backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
pydantic-settings==2.7.1
supabase==2.11.0
openai==1.59.3
httpx==0.28.1
python-multipart==0.0.20
python-jose[cryptography]==3.3.0
websockets==14.1
aiofiles==24.1.0
fpdf2==2.8.1
